requirements.txt
# Code Standards and Typing
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0

# Testing and Mocking
pytest>=8.0.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
pdfplumber>=0.11.0
python-docx>=1.1.0

# Add AI specific libraries here as needed
# openai>=1.12.0
# langchain>=0.1.0
